+++ python/tests/integration/test_config_integration.py
@@ -89,25 +89,6 @@ class TestConfigIntegration:
         assert config.vector_store.table_name == "test_documents"
         assert config.vector_store.similarity_measure == "cosine"
 
-        # Verify LLM provider settings
-        assert config.llm.default_provider == "openai"
-        assert config.llm.embedding_model == "text-embedding-3-large"
-        assert config.llm.openai_api_key == "test-openai-key"
-        assert config.llm.openai_model == "gpt-4"
-        assert config.llm.anthropic_api_key == "test-anthropic-key"
-        assert config.llm.anthropic_model == "claude-3-sonnet"
-
-        # Verify agent configuration
-        assert "test-agent" in config.agents
-        agent = config.agents["test-agent"]
-        assert agent.name == "Test Agent"
-        assert agent.description == "Integration test agent"
-        assert len(agent.sources) == 2
-        assert agent.max_source_count == 5
-        assert agent.similarity_threshold == 0.5
-        assert agent.contract_template == "Test contract template"
-        assert agent.test_template == "Test template"
-
     def test_environment_override_integration(
         self,
         sample_config_file: Path,
@@ -127,26 +108,9 @@ class TestConfigIntegration:
         assert config.vector_store.host == "env-override-host"
         assert config.vector_store.port == 6543
         assert config.vector_store.password == "env-password"
-        assert config.llm.openai_api_key == "env-openai-key"
-        assert config.llm.anthropic_api_key == "env-anthropic-key"
 
         # Check non-overridden values remain
         assert config.vector_store.database == "test_cairo"
-        assert config.llm.openai_model == "gpt-4"
-
-    def test_validation_integration(self, sample_config_file: Path) -> None:
-        """Test configuration validation with real config."""
-        config = ConfigManager.load_config(sample_config_file)
-
-        # Should pass validation with valid config
-        ConfigManager.validate_config(config)
-
-        # Test validation failures
-        config.llm.openai_api_key = None
-        config.llm.anthropic_api_key = None
-        config.llm.gemini_api_key = None
-        with pytest.raises(ValueError, match="At least one LLM provider"):
-            ConfigManager.validate_config(config)
 
     def test_dsn_generation(self, sample_config_file: Path) -> None:
         """Test PostgreSQL DSN generation."""
@@ -154,22 +118,6 @@ class TestConfigIntegration:
 
         expected_dsn = "postgresql://test_user:test_password@test-db.example.com:5433/test_cairo"
         assert config.vector_store.dsn == expected_dsn
-
-    def test_agent_retrieval(self, sample_config_file: Path) -> None:
-        """Test agent configuration retrieval."""
-        config = ConfigManager.load_config(sample_config_file)
-
-        # Get custom agent
-        agent = ConfigManager.get_agent_config(config, "test-agent")
-        assert agent.name == "Test Agent"
-
-        # Get default agent (should exist from Config.__post_init__)
-        cairo_coder_agent = ConfigManager.get_agent_config(config, 'cairo-coder')
-        assert cairo_coder_agent.id == "cairo-coder"
-
-        # Try to get non-existent agent
-        with pytest.raises(ValueError, match="Agent 'unknown' not found"):
-            ConfigManager.get_agent_config(config, "unknown")
 
     @pytest.mark.asyncio
     async def test_missing_config_file(self) -> None:
